requirements.txt
azure-core==1.29.2
azure-search-documents==11.6.0b3
openai==1.25.1
langchain==0.2.5
pypdf==4.2.0
azure-identity==1.16.1
tiktoken==0.7.0
gradio==4.36.1